fix: extract shared test infrastructure into RimfrostTestSupport base class
